Remove the package com.bdt.somepackage and replace it with com.bdt.kiradb.
Add a class Core.java in the new com.bdt.kiradb.  The class contains
nothing yet, and change the name to suit your tastes.  But git will
not track and empty directory, so I added the class (and a counterpart
in src/test/java) to anchor the new package directory com.bdt.kiradb.