Implemented FileBackingStore BackingStore and created a test for it in CoreTest
The test does not remove the left over "person" directory yet.

The S3BackingStore is implemented but untested.  We don't want
to include AWS credentials in the distro so we would want to h
have the user create a S3.properties type file with the AWS setup
and perhaps have the S3 ests look for that properties file and
warn (skip the test) if it isn't there.  Other ideas?