Add a zero arg constructor for S3KiraDb, which assumes an /s3.properties file is available on the classpath.  That properties file is of the form
aws.key=
aws.secret=
aws.bucket=

If the zero arg constructor is used and no /s3.properties is found
on the classpath, or one or more entries in the backing Properties
is of zero length, then a RuntimeException is thrown.  This implies,
for both methods of constructing an S3KiraDb, that a fully functional
S3 store must be configurable at the time of construction.  Otherwise
the code fails fast in both ctors.

The test method
com.bdt.kiradb.S3BackingStoreTest#testS3BackingStoreWithS3Properties is
code to expect an exception because by default /s3.properties exists,
but contains no valid set of S3 credentials.